refactor: move to_pcf into dedicated submodule
